api/requirements.txt
fastapi==0.114.2
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-multipart==0.0.9
weaviate-client==4.7.0
pypdf==4.3.1
tiktoken==0.7.0
httpx==0.27.0
orjson==3.10.7
numpy==1.26.4
pytest==8.3.3
reportlab==4.2.2
python-dotenv==1.0.1


